so small progress on bulk stuff
