+++ source/SIL.AppBuilder.Portal.Frontend/package.json
@@ -67,6 +67,7 @@
     "react-state-helpers": "^1.6.6",
     "react-timezone": "^2.1.0",
     "react-toastify": "^4.5.2",
+    "react-use": "^9.0.0",
     "react-useinterval": "^1.0.1",
     "recompose": "^0.30.0",
     "redux": "^4.0.0",
