+++ source/SIL.AppBuilder.Portal.Frontend/src/ui/routes/projects/new/display.tsx
@@ -42,12 +42,12 @@ export default function Display({ currentOrganizationId, currentOrganization, cr
     try {
       const project = await create({ name, language, isPublic, description }, groupId, typeId);
       const id = idFromRecordIdentity(dataStore, project);
+      
       history.push(`/projects/${id}`);
     } catch (e) {
       toast.error(e);
+      setDisableSubmit(false);
     }
-
-    setDisableSubmit(false);
   }, [name, language, isPublic, description, groupId, typeId]);
 
   const submitClasses = `
